fix: CORS 오류 수정 및 main.ts 정리
- CORS 때문에 Swagger를 이용하지 못하는 오류 수정
- main.ts의 설정 값을 Config Service로 이동
